move provider to contracts
