+++ lib/contracts/provider.js
@@ -2,7 +2,7 @@ const ProviderEngine = require('web3-provider-engine');
 const RpcSubprovider = require('web3-provider-engine/subproviders/rpc.js');
 const bip39 = require("bip39");
 const hdkey = require('ethereumjs-wallet/hdkey');
-const fs = require('./fs');
+const fs = require('../core/fs');
 
 class Provider {
   constructor(options) {
