+++ lib/core/engine.js
@@ -10,7 +10,7 @@ const Watch = require('../pipeline/watch.js');
 const LibraryManager = require('../versions/library_manager.js');
 const Pipeline = require('../pipeline/pipeline.js');
 const async = require('async');
-const Provider = require('./provider');
+const Provider = require('../contracts/provider');
 
 class Engine {
   constructor(options) {
